Correction code final
-ajout methode conversion , en .pour requete insert
-implementation de la methode finalpour partager les donnée

+++ cpp/josh/codeFinal/Anemometre.h
@@ -8,7 +8,7 @@ class Anemometre : public Capteur
 {
 
     public:
-        Anemometre(double channel);
+        Anemometre(int channel);
 		double readValue();
 		string getUnit();
